[macsec testcase] Split macsec testcase, refine execution speed with multi-thread (#6641)
* Split testcases

* Polish code

* Refine execution speed with multi-thread, add more logs

* Fix process stuck issue, use self-hosted process pool instead of
official thread pool

* Manually cherry-pick #6732
