+++ tests/kvmtest.sh
@@ -202,7 +202,7 @@ test_t0_sonic() {
     tgname=t0-sonic
     tests="\
       bgp/test_bgp_fact.py \
-      macsec/test_macsec.py"
+      macsec"
 
     pushd $SONIC_MGMT_DIR/tests
     ./run_tests.sh $RUNTEST_CLI_COMMON_OPTS -c "$tests" -p logs/$tgname -e "--neighbor_type=sonic --enable_macsec --macsec_profile=128_SCI,256_XPN_SCI"
